Move content of src/pharmpy/model/results.py to src/pharmpy/workflows/results.py
