+++ src/pharmpy/model/__init__.py
@@ -5,7 +5,6 @@ from .estimation import EstimationStep, EstimationSteps
 from .model import Model, ModelError, ModelfitResultsError, ModelSyntaxError
 from .parameters import Parameter, Parameters
 from .random_variables import RandomVariables, VariabilityHierarchy, VariabilityLevel
-from .results import Results
 from .statements import (
     Assignment,
     Bolus,
@@ -43,7 +42,6 @@ __all__ = (
     'Parameter',
     'Parameters',
     'RandomVariables',
-    'Results',
     'Statement',
     'Statements',
     'to_compartmental_system',
