+++ src/pharmpy/tools/bootstrap/results.py
@@ -7,10 +7,10 @@ import pharmpy.visualization
 from pharmpy.deps import numpy as np
 from pharmpy.deps import pandas as pd
 from pharmpy.deps.scipy import stats
-from pharmpy.model import Model, Results
+from pharmpy.model import Model
 from pharmpy.tools import read_modelfit_results
 from pharmpy.tools.psn_helpers import cmd_line_model_path, model_paths
-from pharmpy.workflows.results import ModelfitResults
+from pharmpy.workflows.results import ModelfitResults, Results
 
 
 @dataclass(frozen=True)
