+++ src/pharmpy/tools/run.py
@@ -13,12 +13,12 @@ import pharmpy.workflows.results
 from pharmpy.deps import numpy as np
 from pharmpy.deps import pandas as pd
 from pharmpy.internals.fs.path import normalize_user_given_path
-from pharmpy.model import Model, Results
+from pharmpy.model import Model
 from pharmpy.modeling import calculate_aic, calculate_bic, check_high_correlations, read_model
 from pharmpy.modeling.lrt import degrees_of_freedom as lrt_df
 from pharmpy.modeling.lrt import test as lrt_test
 from pharmpy.tools.psn_helpers import create_results as psn_create_results
-from pharmpy.workflows import Workflow, execute_workflow, split_common_options
+from pharmpy.workflows import Results, Workflow, execute_workflow, split_common_options
 from pharmpy.workflows.model_database import LocalModelDirectoryDatabase, ModelDatabase
 from pharmpy.workflows.results import ModelfitResults, mfr
 from pharmpy.workflows.tool_database import ToolDatabase
