+++ src/pharmpy/tools/simeval/results.py
@@ -5,10 +5,11 @@ from pathlib import Path
 from typing import Any, Optional
 
 from pharmpy.deps import pandas as pd
-from pharmpy.model import Model, Results
+from pharmpy.model import Model
 from pharmpy.modeling import plot_individual_predictions
 from pharmpy.tools import read_modelfit_results
 from pharmpy.tools.simfit.results import psn_simfit_results
+from pharmpy.workflows import Results
 
 
 @dataclass(frozen=True)
