+++ src/pharmpy/workflows/__init__.py
@@ -37,6 +37,7 @@ from .model_database import (
     ModelDatabase,
     NullModelDatabase,
 )
+from .results import Results
 from .task import Task
 from .tool_database import LocalDirectoryToolDatabase, NullToolDatabase, ToolDatabase
 from .workflow import Workflow, WorkflowBuilder
@@ -88,6 +89,7 @@ __all__ = [
     'NullModelDatabase',
     'NullToolDatabase',
     'ModelDatabase',
+    'Results',
     'Task',
     'ToolDatabase',
     'Workflow',
