+++ tox.ini
@@ -143,7 +143,7 @@ deps =
     pytest>5.3.5
     pyright
 commands =
-    pyright {posargs:src/pharmpy/deps src/pharmpy/internals src/pharmpy/model/parameters.py src/pharmpy/model/distributions src/pharmpy/model/random_variables.py src/pharmpy/model/__init__.py src/pharmpy/model/data.py src/pharmpy/model/datainfo.py src/pharmpy/model/estimation.py src/pharmpy/model/results.py}
+    pyright {posargs:src/pharmpy/deps src/pharmpy/internals src/pharmpy/model/parameters.py src/pharmpy/model/distributions src/pharmpy/model/random_variables.py src/pharmpy/model/__init__.py src/pharmpy/model/data.py src/pharmpy/model/datainfo.py src/pharmpy/model/estimation.py}
 
 [testenv:{py39-,py310-,py311-,py312-,}report]
 deps = coverage
